Use backed mode to read H5AD when validating

Primary change: significantly reduce memory use when validating CSR-formatted data. The `X` and `raw.X` are read incrementally, in chunks, dramatically reducing memory usage. Dense data is run in backed mode, and should also see reduced memory use. CSC is still loaded into memory as before, and has no change.

Changes:
* revise Validator() to use backed mode when reading AnnData
* incrementally compute over `X` and `raw.X`
* reformat validate.py to conform with repo Python style configuration
* clean up debt in tests related to Pandas and AnnData

Old and new code was compared:
* there is a slight performance penalty (less than 10%) for using backed mode on small to mid-sized datasets
* large CSR datasets now complete without an OOM failure (or excessive swapping).  This was tested to datasets significantly exceeding the size of main memory (eg., a 500GiB+ H5AD on a 64GiB machine with no swap).

+++ cellxgene_schema_cli/tests/test_validate.py
@@ -250,7 +250,7 @@ class TestSeuratConvertibility(unittest.TestCase):
     def test_determine_seurat_convertibility(self):
         # Sparse matrix with too many nonzero values is not Seurat-convertible
         sparse_matrix_too_large = sparse.csr_matrix(
-            np.ones((good_obs.shape[0], good_var.shape[0]))
+            np.ones((good_obs.shape[0], good_var.shape[0]), dtype=np.float32)
         )
         self.validation_helper(sparse_matrix_too_large)
         self.validator._validate_seurat_convertibility()
@@ -259,7 +259,7 @@ class TestSeuratConvertibility(unittest.TestCase):
 
         # Reducing nonzero count by 1, to within limit, makes it Seurat-convertible
         sparse_matrix_with_zero = sparse.csr_matrix(
-            np.ones((good_obs.shape[0], good_var.shape[0]))
+            np.ones((good_obs.shape[0], good_var.shape[0]), dtype=np.float32)
         )
         sparse_matrix_with_zero[0, 0] = 0
         self.validation_helper(sparse_matrix_with_zero)
@@ -268,7 +268,9 @@ class TestSeuratConvertibility(unittest.TestCase):
         self.assertTrue(self.validator.is_seurat_convertible)
 
         # Dense matrices with a dimension that exceeds limit will fail -- zeros are irrelevant
-        dense_matrix_with_zero = np.zeros((good_obs.shape[0], good_var.shape[0]))
+        dense_matrix_with_zero = np.zeros(
+            (good_obs.shape[0], good_var.shape[0]), dtype=np.float32
+        )
         self.validation_helper(dense_matrix_with_zero)
         self.validator.schema_def["max_size_for_seurat"] = 2**2 - 1
         self.validator._validate_seurat_convertibility()
@@ -276,7 +278,7 @@ class TestSeuratConvertibility(unittest.TestCase):
         self.assertFalse(self.validator.is_seurat_convertible)
 
         # Dense matrices with dimensions in bounds but total count over will succeed
-        dense_matrix = np.ones((good_obs.shape[0], good_var.shape[0]))
+        dense_matrix = np.ones((good_obs.shape[0], good_var.shape[0]), dtype=np.float32)
         self.validation_helper(dense_matrix)
         self.validator.schema_def["max_size_for_seurat"] = 2**3 - 1
         self.validator._validate_seurat_convertibility()
